Package structure changed. The type property of response classes is set with @JsonTypeName instead of @JsonSubTypes. Notification to notification response converters return business domain specific classes.
