+++ src/main/java/io/knowledgeassets/myskills/server/communityuser/UserKickedOutFromCommunityNotificationConverter.java
@@ -1,4 +1,4 @@
-package io.knowledgeassets.myskills.server.communityuser.notification.kickout;
+package io.knowledgeassets.myskills.server.communityuser;
 
 import io.knowledgeassets.myskills.server.notification.AbstractNotificationResponse;
 import org.springframework.core.convert.converter.Converter;
