+++ src/main/java/io/knowledgeassets/myskills/server/communityuser/UserLeftCommunityNotification.java
@@ -1,4 +1,4 @@
-package io.knowledgeassets.myskills.server.communityuser.notification.leaving;
+package io.knowledgeassets.myskills.server.communityuser;
 
 import io.knowledgeassets.myskills.server.community.Community;
 import io.knowledgeassets.myskills.server.notification.Notification;
