+++ src/main/java/io/knowledgeassets/myskills/server/communityuser/command/CommunityUserCommandService.java
@@ -4,8 +4,8 @@ import io.knowledgeassets.myskills.server.community.Community;
 import io.knowledgeassets.myskills.server.community.CommunityRole;
 import io.knowledgeassets.myskills.server.communityuser.CommunityUser;
 import io.knowledgeassets.myskills.server.communityuser.CommunityUserRepository;
-import io.knowledgeassets.myskills.server.communityuser.notification.kickout.UserKickedOutFromCommunityNotification;
-import io.knowledgeassets.myskills.server.communityuser.notification.leaving.UserLeftCommunityNotification;
+import io.knowledgeassets.myskills.server.communityuser.UserKickedOutFromCommunityNotification;
+import io.knowledgeassets.myskills.server.communityuser.UserLeftCommunityNotification;
 import io.knowledgeassets.myskills.server.exception.DuplicateResourceException;
 import io.knowledgeassets.myskills.server.exception.NoSuchResourceException;
 import io.knowledgeassets.myskills.server.exception.enums.Model;
