+++ src/main/java/io/knowledgeassets/myskills/server/communityuser/registration/AcceptanceToCommunityNotificationConverter.java
@@ -1,4 +1,4 @@
-package io.knowledgeassets.myskills.server.communityuser.registration.notification.acceptance;
+package io.knowledgeassets.myskills.server.communityuser.registration;
 
 import io.knowledgeassets.myskills.server.notification.AbstractNotificationResponse;
 import org.springframework.core.convert.converter.Converter;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 public class AcceptanceToCommunityNotificationConverter implements Converter<AcceptanceToCommunityNotification, AbstractNotificationResponse> {
 
 	@Override
-	public AbstractNotificationResponse convert(AcceptanceToCommunityNotification notification) {
+	public AcceptanceToCommunityNotificationResponse convert(AcceptanceToCommunityNotification notification) {
 		return AcceptanceToCommunityNotificationResponse.of(notification);
 	}
 }
