+++ src/main/java/io/knowledgeassets/myskills/server/communityuser/registration/InvitationToJoinCommunityNotification.java
@@ -1,4 +1,4 @@
-package io.knowledgeassets.myskills.server.communityuser.registration.notification.acceptance;
+package io.knowledgeassets.myskills.server.communityuser.registration;
 
 import io.knowledgeassets.myskills.server.communityuser.registration.CommunityUserRegistration;
 import io.knowledgeassets.myskills.server.notification.Notification;
@@ -13,13 +13,13 @@ import java.time.LocalDateTime;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class AcceptanceToCommunityNotification extends Notification {
+public class InvitationToJoinCommunityNotification extends Notification {
 
 	@Relationship(type = "CAUSED_BY")
 	private CommunityUserRegistration registration;
 
 	@Builder
-	public AcceptanceToCommunityNotification(String id, LocalDateTime creationDatetime, CommunityUserRegistration registration) {
+	public InvitationToJoinCommunityNotification(String id, LocalDateTime creationDatetime, CommunityUserRegistration registration) {
 		super(id, creationDatetime);
 		this.registration = registration;
 	}
