+++ src/main/java/io/knowledgeassets/myskills/server/communityuser/registration/RequestToJoinCommunityNotificationConverter.java
@@ -1,4 +1,4 @@
-package io.knowledgeassets.myskills.server.communityuser.registration.notification.request;
+package io.knowledgeassets.myskills.server.communityuser.registration;
 
 import io.knowledgeassets.myskills.server.notification.AbstractNotificationResponse;
 import org.springframework.core.convert.converter.Converter;
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Component;
 public class RequestToJoinCommunityNotificationConverter implements Converter<RequestToJoinCommunityNotification, AbstractNotificationResponse> {
 
 	@Override
-	public AbstractNotificationResponse convert(RequestToJoinCommunityNotification notification) {
+	public RequestToJoinCommunityNotificationResponse convert(RequestToJoinCommunityNotification notification) {
 		return RequestToJoinCommunityNotificationResponse.of(notification);
 	}
 }
