+++ src/main/java/io/knowledgeassets/myskills/server/communityuser/registration/command/CommunityUserRegistrationCommandService.java
@@ -3,9 +3,9 @@ package io.knowledgeassets.myskills.server.communityuser.registration.command;
 import io.knowledgeassets.myskills.server.community.Community;
 import io.knowledgeassets.myskills.server.community.CommunityRole;
 import io.knowledgeassets.myskills.server.communityuser.command.CommunityUserCommandService;
-import io.knowledgeassets.myskills.server.communityuser.registration.notification.acceptance.AcceptanceToCommunityNotification;
-import io.knowledgeassets.myskills.server.communityuser.registration.notification.invitation.InvitationToJoinCommunityNotification;
-import io.knowledgeassets.myskills.server.communityuser.registration.notification.request.RequestToJoinCommunityNotification;
+import io.knowledgeassets.myskills.server.communityuser.registration.AcceptanceToCommunityNotification;
+import io.knowledgeassets.myskills.server.communityuser.registration.InvitationToJoinCommunityNotification;
+import io.knowledgeassets.myskills.server.communityuser.registration.RequestToJoinCommunityNotification;
 import io.knowledgeassets.myskills.server.notification.command.NotificationCommandService;
 import io.knowledgeassets.myskills.server.security.CurrentUserService;
 import io.knowledgeassets.myskills.server.user.User;
