+++ src/main/java/io/knowledgeassets/myskills/server/notification/AbstractNotificationResponse.java
@@ -1,11 +1,6 @@
 package io.knowledgeassets.myskills.server.notification;
 
-import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import io.knowledgeassets.myskills.server.communityuser.registration.notification.acceptance.AcceptanceToCommunityNotificationResponse;
-import io.knowledgeassets.myskills.server.communityuser.registration.notification.invitation.InvitationToJoinCommunityNotificationResponse;
-import io.knowledgeassets.myskills.server.communityuser.registration.notification.request.RequestToJoinCommunityNotificationResponse;
-import io.knowledgeassets.myskills.server.communityuser.notification.kickout.UserKickedOutFromCommunityNotificationResponse;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
@@ -14,12 +9,6 @@ import java.time.LocalDateTime;
 
 @Data
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-@JsonSubTypes({
-		@JsonSubTypes.Type(value = RequestToJoinCommunityNotificationResponse.class),
-		@JsonSubTypes.Type(value = InvitationToJoinCommunityNotificationResponse.class),
-		@JsonSubTypes.Type(value = AcceptanceToCommunityNotificationResponse.class),
-		@JsonSubTypes.Type(value = UserKickedOutFromCommunityNotificationResponse.class)
-})
 @ApiModel(
 		value = "AbstractNotificationResponse",
 		description = "This holds notification data. It is used to transfer notification data to a client. The field \"type\" is not documented here as it is added by Jackson automatically. " +
