+++ src/test/java/io/knowledgeassets/myskills/server/notification/command/NotificationCommandServiceTests.java
@@ -3,7 +3,7 @@ package io.knowledgeassets.myskills.server.notification.command;
 import io.knowledgeassets.myskills.server.community.Community;
 import io.knowledgeassets.myskills.server.community.CommunityType;
 import io.knowledgeassets.myskills.server.communityuser.registration.CommunityUserRegistration;
-import io.knowledgeassets.myskills.server.communityuser.registration.notification.invitation.InvitationToJoinCommunityNotification;
+import io.knowledgeassets.myskills.server.communityuser.registration.InvitationToJoinCommunityNotification;
 import io.knowledgeassets.myskills.server.notification.Notification;
 import io.knowledgeassets.myskills.server.notification.NotificationRepository;
 import io.knowledgeassets.myskills.server.user.User;
